requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
browser-cookie3==0.19.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
curl-cffi==0.5.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
g4f==0.0.2.4
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
Js2Py==0.74
lz4==4.3.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
pluggy==1.3.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pyjsparser==2.7.1
pytest==7.4.0
python-dotenv==1.0.0
quickjs==1.19.2
regex==2023.8.8
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.21
tiktoken==0.4.0
typing_extensions==4.8.0
tzlocal==5.0.1
urllib3==2.0.4
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
